Update ProblemTest.java
It's best to use more descriptive test names so that upon seeing the test file for the first time, people can guess what each test case is actually testing. Also, for real code, it's good to try to test for corner cases so you know what your code does with null pointers for example. (In this case I think it'll crash as is... Which should either (1) be fixed or (2) be documented so users know what to expect.)